refactor: split SecurityAssessor and ToolAnnotationAssessor into focused modules
SecurityAssessor.ts (2,165 → 326 lines):
- Extract SecurityResponseAnalyzer (~1,100 lines) - response analysis, reflection detection
- Extract SecurityPayloadTester (~530 lines) - test execution, batching, progress
- Extract SecurityPayloadGenerator (~170 lines) - parameter creation, tool classification
- Create securityTests/index.ts barrel export

ToolAnnotationAssessor.ts (1,856 → 1,110 lines):
- Extract DescriptionPoisoningDetector (~330 lines) - poisoning patterns, detection
- Extract AnnotationDeceptionDetector (~210 lines) - keyword arrays, deception logic
- Extract BehaviorInference (~200 lines) - behavior inference from tool patterns
- Create annotations/index.ts barrel export

All 2,325 assessment tests passing. Build verified.
