+++ client/src/services/assessment/modules/securityTests/index.ts
@@ -1,0 +1,20 @@
+/**
+ * Security Assessment Module
+ * Exports all security-related components
+ */
+
+export {
+  SecurityResponseAnalyzer,
+  type ConfidenceResult,
+  type AnalysisResult,
+  type ErrorClassification,
+} from "./SecurityResponseAnalyzer";
+
+export {
+  SecurityPayloadTester,
+  type TestProgressCallback,
+  type PayloadTestConfig,
+  type TestLogger,
+} from "./SecurityPayloadTester";
+
+export { SecurityPayloadGenerator } from "./SecurityPayloadGenerator";
